requirements.txt
fastapi
uvicorn[standard]
jsonschema
psycopg2-binary
pytest